+++ database/migrations/0000_00_00_000007_create_tags_table.php
@@ -5,31 +5,29 @@ use Illuminate\Database\Schema\Blueprint;
 
 class CreateTagsTable extends Migration {
 
-	/**
-	 * Run the migrations.
-	 *
-	 * @return void
-	 */
-	public function up()
-	{
-		Schema::create('tags', function(Blueprint $table)
-		{
-			$table->integer('id_post')->unsigned();
-			$table->foreign('id_post')->references('id')->on('posts')->onUpdate('RESTRICT')->onDelete('CASCADE');
-			$table->string('tag', 255);
-			$table->timestamps();
-		});
-	}
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up() {
+        if (!Schema::hasTable('tags')) {
+            Schema::create('tags', function(Blueprint $table) {
+                $table->integer('id_post')->unsigned();
+                $table->foreign('id_post')->references('id')->on('posts')->onUpdate('RESTRICT')->onDelete('CASCADE');
+                $table->string('tag', 255);
+                $table->timestamps();
+            });
+        }
+    }
 
-
-	/**
-	 * Reverse the migrations.
-	 *
-	 * @return void
-	 */
-	public function down()
-	{
-		Schema::drop('tags');
-	}
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down() {
+        Schema::dropIfExists('tags');
+    }
 
 }
